+++ scenario/scenario.go
@@ -60,25 +60,6 @@ func defaultTestOrder() *zeroex.Order {
 	}
 }
 
-func defaultTestOrderV4() *zeroex.OrderV4 {
-	return &zeroex.OrderV4{
-		ChainID:             big.NewInt(constants.TestChainID),
-		ExchangeAddress:     ganacheAddresses.ExchangeProxy,
-		MakerToken:          constants.NullAddress,
-		TakerToken:          constants.NullAddress,
-		MakerAmount:         big.NewInt(100),
-		TakerAmount:         big.NewInt(42),
-		TakerTokenFeeAmount: big.NewInt(0),
-		Maker:               constants.GanacheAccount1,
-		Taker:               constants.NullAddress,
-		Sender:              constants.NullAddress,
-		FeeRecipient:        constants.NullAddress,
-		Pool:                zeroex.HexToBytes32("0000000000000000000000000000000000000000000000000000000000000000"),
-		Expiry:              big.NewInt(time.Now().Add(24 * time.Hour).Unix()),
-		Salt:                big.NewInt(int64(time.Now().Nanosecond())),
-	}
-}
-
 func defaultConfig() *orderopts.Config {
 	return &orderopts.Config{
 		Order:             defaultTestOrder(),
